Refactor Navigation
