+++ src/pages/EditorPage/Welcome/index.js
@@ -1,7 +1,13 @@
 import React from "react";
+import AddModalContent from "../Modals/AddModalContent";
 import { ModalTypes } from "../utils/const";
 
 export default function Welcome({ setShowModal, createFile, showEditor }) {
+  const handleNew = (type) => {
+    createFile(type);
+    setShowModal(ModalTypes.RenameModal);
+  };
+
   return (
     <div
       className={`text-center d-flex justify-content-center min-vh-100 ${
@@ -21,25 +27,7 @@ export default function Welcome({ setShowModal, createFile, showEditor }) {
           Use this sandbox to create, inspect, modify, and compose components to
           create new experiences on NEAR.
         </p>
-        <div className="d-flex justify-content-center mt-5">
-          <button
-            className="btn btn-outline-success mb-3"
-            style={{ width: "250px" }}
-            onClick={() => setShowModal(ModalTypes.OpenModal)}
-          >
-            Open Component
-          </button>
-        </div>
-        <div className="w-100 text-center text-secondary mb-3">or</div>
-        <div className="d-flex justify-content-center">
-          <button
-            className="btn btn-success mb-4"
-            style={{ width: "250px" }}
-            onClick={() => (setShowModal(), createFile(Filetype.Widget))}
-          >
-            Create New Component
-          </button>
-        </div>
+        <AddModalContent setShowModal={setShowModal} handleNew={handleNew} />
       </div>
     </div>
   );
